refactor: rename transactionScope to transactionPathParams :hammer:
